hedge_fund_bot/requirements.txt
langchain==0.3.0
langgraph==0.2.0
langchain-groq==0.2.0
langchain-community==0.3.0
yfinance==0.2.43
pandas==2.2.3
duckduckgo-search==3.9.20
python-dotenv==1.0.1
